animationLoginAndCardDnD

+++ src/components/TopBar.jsx
@@ -4,9 +4,10 @@ import { BsChevronDown, BsPlusCircleDotted } from "react-icons/bs";
 import { getProjectUser } from '../api/users';
 import { getProject } from '../api/project';
 import { useQuery } from 'react-query';
-import { useParams, Link, useNavigate } from 'react-router-dom'
+import { useParams, Link, useNavigate, useLocation } from 'react-router-dom'
 import { GrFormClose } from "react-icons/gr";
 import Modal from './Modal';
+import Swal from 'sweetalert2';
 
 export default function TopBar() {
   const [projectUsers, setProjectUsers] = useState([{ id: "", username: "" }]);
@@ -14,6 +15,7 @@ export default function TopBar() {
   const [referralCodeModalOpen, setReferralCodeModalOpen] = useState(false);
   const { projectId } = useParams();
   const navigate = useNavigate();
+  const location = useLocation();
 
   const getProjectUserQuery = useQuery("getProjectUser", () => getProjectUser(projectId),
     {
@@ -40,8 +42,41 @@ export default function TopBar() {
   }
 
   const handleLogout = () => {
-    localStorage.clear();
-    navigate("/");
+    Swal.fire({
+      title: "登出",
+      text: "確定要登出嗎?",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#5BA491",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "確定",
+      cancelButtonText: "取消"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        navigate("/");
+      }
+    });
+
+
+  }
+
+  if (location.pathname === "/homepage") {
+    return (
+      <div className='fixed z-40 h-16 w-full bg-[#FFFFFF] flex items-center justify-between pr-5 border-b-2'>
+        <Link to="/homepage" onClick={cleanStage} className="flex px-5 items-center font-bold font-Mulish text-2xl">
+          <img src="/SDLS_LOGOO.jpg" alt="Logo" className="h-14 w-auto" />
+        </Link>
+        <div className="flex items-center">
+          <h3 className="font-bold cursor-pointer p-1 mr-2 rounded-lg">
+            {localStorage.getItem("username")}
+          </h3>
+          <h3 className="font-bold cursor-pointer p-1 mr-2 hover:bg-slate-100 rounded-lg" onClick={handleLogout}>
+            登出
+          </h3>
+        </div>
+      </div>
+    );
   }
 
   return (
@@ -50,7 +85,7 @@ export default function TopBar() {
       <Link to="/homepage" onClick={cleanStage} className="flex px-5 items-center font-bold font-Mulish text-2xl">
         {/* <img src="/logo.ico" alt="Logo" className="mr-2" /> */}
         <img src="/SDLS_LOGOO.jpg" alt="Logo" className=" h-14 w-auto" />
-        
+
       </Link>
       <div className="flex items-center">
         <ul className="flex items-center justify-center space-x-1">
